requirements.txt
# Deep learning frameworks
torch
torchvision
tensorflow
keras

# Image processing
opencv-python
pillow

# Utilities
numpy
scikit-learn
matplotlib

# Optional for GUI
tk
